Move info and supporting methods to seperate class
